IMGW hydrological data from the IMGW datastore repository

Downloading hourly and sub-hourly (hydrological) data from the telemetric stations available in the danepubliczne.imgw.pl/datastore collection since 2008. Most parameters are collected with 10 minutes interval and thus it is recommended to download only the mandatory years, parameters or stations. For example, 1 year of data with all available parameters requires processing around 2-4GB of uncompressed data.

hydro_imgw_datastore(
  year,
  parameters = NULL,
  stations = NULL,
  coords = TRUE,
  allow_failure = TRUE
)

Arguments

year

numeric vector of years to be downloaded (e.g., 2022:2023)

parameters

character vector describing which parameters to be downloaded. Default NULL means to download all available.

"level" - Water level (operational) (cm)

"level_obs" - Water level (observer) (cm)

"flow" - Water flow rate (operational) (m3/s)

"temp" - Water temperature (observer) (degree C)

stations

character vector with station name(s) as visible in the hydro_imgw_telemetry_stations(). Default NULL means to download data for all available stations.

coords

logical - whether to append the dataset with station full name, longitude, latitude and altitude. Default: TRUE

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

Value

data.table with a raw hydrorological measurements in 10-min or 60-min intervals. Please note that this dataset is not validated by experts and may contain invalid values.

Details

Data from the IMGW automated (telemetry) systems are non validated by experts and may contain invalid values.

Examples

# \donttest{
imgw_hydro_telemetry = hydro_imgw_datastore(year = 2022,
                                            parameters = "flow",
                                            stations = "FORDON",
                                            coords = TRUE)
# }