Daily IMGW meteorological data

Downloading daily (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo_imgw_daily(
  rank = "synop",
  year,
  status = FALSE,
  coords = FALSE,
  station = NULL,
  col_names = "short",
  allow_failure = TRUE,
  ...
)

Arguments

rank

rank of the stations: "synop" (default), "climate", or "precip"

year

vector of years (e.g., 1966:2000)

status

leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords

add coordinates of the station (logical value TRUE or FALSE)

station

name of meteorological station(s). It accepts names (characters in CAPITAL LETTERS); stations' IDs (numeric) are no longer valid

col_names

three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

...

other parameters that may be passed to the 'shortening' function that shortens column names

Examples

# \donttest{
  daily = meteo_imgw_daily(rank = "climate", year = 2000)
#>     Your system locale is: C.UTF-8 which may cause trouble.
#>     Please consider changing it manually while working with climate, e.g.:
#>     Sys.setlocale(category = 'LC_ALL', locale = 'en_US.UTF-8') 
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/k_d_format.txt"
#> /tmp/Rtmpvpf0le/file15a421938130
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/k_d_t_format.txt"
#> /tmp/Rtmpvpf0le/file15a451ad284d
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/"
#> /tmp/Rtmpvpf0le/file15a453ae5f9c
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/"
#> /tmp/Rtmpvpf0le/file15a453ae5f9c
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/1996_k.zip"
#> /tmp/Rtmpvpf0le/file15a45c3fe09b
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/1997_k.zip"
#> /tmp/Rtmpvpf0le/file15a41b32cf93
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/1998_k.zip"
#> /tmp/Rtmpvpf0le/file15a417b0a27
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/1999_k.zip"
#> /tmp/Rtmpvpf0le/file15a4764acbe6
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/dobowe/klimat/1996_2000/2000_k.zip"
#> /tmp/Rtmpvpf0le/file15a43120069
# }