CO2 Mauna Loa (NOAA) dataset

Carbon Dioxide (CO2) monthly measurements from Mauna Loa observatory. The source file is available at: ftp://aftp.cmdl.noaa.gov/products/trends/co2/co2_mm_mlo.txt with all further details.

meteo_noaa_co2()

Value

Data frame with historical CO2 concentrations

Details

Data from March 1958 through April 1974 have been obtained by C. David Keeling of the Scripps Institution of Oceanography (SIO) and were obtained from the Scripps website (scrippsco2.ucsd.edu).

The "average" column contains the monthly mean CO2 mole fraction determined from daily averages. The mole fraction of CO2, expressed as parts per million (ppm) is the number of molecules of CO2 in every one million molecules of dried air (water vapor removed). If there are missing days concentrated either early or late in the month, the monthly mean is corrected to the middle of the month using the average seasonal cycle. Missing months are denoted by -99.99. The "interpolated" column includes average values from the preceding column and interpolated values where data are missing. Interpolated values are computed in two steps. First, we compute for each month the average seasonal cycle in a 7-year window around each monthly value. In this way the seasonal cycle is allowed to change slowly over time. We then determine the "trend" value for each month by removing the seasonal cycle; this result is shown in the "trend" column. Trend values are linearly interpolated for missing months. The interpolated monthly mean is then the sum of the average seasonal cycle value and the trend value for the missing month. NOTE: In general, the data presented for the last year are subject to change, depending on recalibration of the reference gas mixtures used, and other quality control procedures. Occasionally, earlier years may also be changed for the same reasons. Usually these changes are minor. CO2 expressed as a mole fraction in dry air, micromol/mol, abbreviated as ppm

Examples

# \donttest{
  co2 = meteo_noaa_co2()
#> /tmp/Rtmp95w7yv/file17d079bd8bef
  head(co2)
#>     yy mm     yy_d co2_avg co2_interp co2_seas ndays st_dev_days
#> 1 1958  3 1958.203  315.71     314.44       -1    NA          NA
#> 2 1958  4 1958.288  317.45     315.16       -1    NA          NA
#> 3 1958  5 1958.370  317.51     314.69       -1    NA          NA
#> 4 1958  6 1958.455  317.27     315.15       -1    NA          NA
#> 5 1958  7 1958.537  315.87     315.20       -1    NA          NA
#> 6 1958  8 1958.622  314.93     316.21       -1    NA          NA
# }