Hydrological data from IMGW

Downloading daily, and monthly hydrological data from the measurement stations available in the danepubliczne.imgw.pl collection

hydro_imgw(
  interval,
  year,
  coords = FALSE,
  value = "H",
  station = NULL,
  col_names = "short",
  ...
)

Arguments

interval

temporal resolution of the data ("daily" , "monthly", or "semiannual_and_annual")

year

vector of years (e.g., 1966:2000)

coords

add coordinates of the stations (logical value TRUE or FALSE)

value

type of data (can be: state - "H" (default), flow - "Q", or temperature - "T")

station

vector of hydrological stations danepubliczne.imgw.pl; can be given as station name with CAPITAL LETTERS (character) It accepts either names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names

three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...

other parameters that may be passed to the 'shortening' function that shortens column names

Value

A data.frame with columns describing the hydrological parameters (e.g. flow, water level) where each row represent a measurement, depending on the interval, at a given hour, month or year. If coords = TRUE additional two columns with geografic coordinates are added.

Examples

# \donttest{
  x = hydro_imgw("monthly", year = 1999)
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_hydrologiczne/miesieczne/"
#> /tmp/RtmpRYa28L/file1827369510f
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_hydrologiczne/miesieczne/mies_info.txt"
#> /tmp/RtmpRYa28L/file182516decdf
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_hydrologiczne/miesieczne/1999/mies_1999.zip"
#> /tmp/RtmpRYa28L/file182eb15b2d
  head(x)
#>              id station riv_or_lake  hyy idhyy idex   H   Q  T mm
#> 11665 150210180 ANNOPOL   Wisła (2) 1999     1    1 249 250 NA 11
#> 11666 150210180 ANNOPOL   Wisła (2) 1999     1    2 291 413 NA 11
#> 11667 150210180 ANNOPOL   Wisła (2) 1999     1    3 325 572 NA 11
#> 11668 150210180 ANNOPOL   Wisła (2) 1999     2    1 232 191 NA 12
#> 11669 150210180 ANNOPOL   Wisła (2) 1999     2    2 271 341 NA 12
#> 11670 150210180 ANNOPOL   Wisła (2) 1999     2    3 384 904 NA 12
# }