Meteorological data from IMGW

Downloading hourly, daily, and monthly meteorological data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo_imgw(
  interval,
  rank = "synop",
  year,
  status = FALSE,
  coords = FALSE,
  station = NULL,
  col_names = "short",
  ...
)

Arguments

interval

temporal resolution of the data ("hourly", "daily", "monthly")

rank

rank of the stations: "synop" (default), "climate" or "precip"

year

vector of years (e.g., 1966:2000)

status

leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords

add coordinates of the station (logical value TRUE or FALSE)

station

vector of hydrological stations danepubliczne.imgw.pl can be name of station CAPITAL LETTERS(character). It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names

three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...

other parameters that may be passed to the 'shortening' function that shortens column names

Value

A data.frame with columns describing the meteorological parameters (e.g. temperature, wind speed, precipitation) where each row represent a measurement, depending on the interval, at a given hour, month or year. If coords = TRUE additional two columns with geographic coordinates are added.

Examples

# \donttest{
  x = meteo_imgw("monthly", year = 2018, coords = TRUE)
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/s_m_d_format.txt"
#> /tmp/RtmpAol0e5/file18060fd2572
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/s_m_t_format.txt"
#> /tmp/RtmpAol0e5/file180234af67c
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/"
#> /tmp/RtmpAol0e5/file180669cd17e
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/2018/2018_m_s.zip"
#> /tmp/RtmpAol0e5/file1803f46f3a8
  head(x)
#>           rank        id        X        Y       station   yy mm tmax_abs
#> 12 SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  1     11.6
#> 11 SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  2      7.2
#> 4  SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  3     16.5
#> 5  SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  4     26.7
#> 9  SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  5     28.0
#> 6  SYNOPTYCZNA 349190600 19.00234 49.80671 BIELSKO-BIAŁA 2018  6     29.5
#>    tmax_mean tmin_abs tmin_mean t2m_mean_mon t5cm_min rr_monthly rr_max_daily
#> 12       5.0     -7.7      -0.4          2.3    -10.8       23.5          6.5
#> 11      -1.4    -15.5      -6.0         -3.7    -23.3       34.1          9.5
#> 4        4.8    -17.2      -3.0          1.0    -21.9       28.5         10.1
#> 5       20.3     -1.0       8.2         14.3     -1.7       10.3          3.5
#> 9       21.3      5.1      10.6         16.2      2.2       91.1         19.5
#> 6       22.7      8.1      13.5         17.6      6.1      135.5         30.4
#>    first_day_max_rr last_day_max_rr insolation_monthly snowcover_max
#> 12                8              NA               58.8             4
#> 11                4              NA               47.0            19
#> 4                16              NA              110.7            15
#> 5                23              NA              218.2             0
#> 9                17              NA              247.1             0
#> 6                 2              NA              164.1             0
#>    snowcover_days rain_days snow_days r_s_days hail_days fog_days
#> 12              5         7        10        4         0        5
#> 11             25         0        16        4         0        7
#> 4              14         5         7        2         0        2
#> 5               0         7         1        1         0        1
#> 9               0        14         0        0         1        1
#> 6               0        16         1        0         0        2
#>    fogginess_days rime_days glaze_days snowstorm_low_days snowstorm_high_days
#> 12             25         0          0                  0                   0
#> 11             28         0          1                  6                   1
#> 4              24         0          0                  2                   0
#> 5               6         0          0                  0                   0
#> 9              11         0          0                  0                   0
#> 6              15         0          0                  0                   0
#>    hazyness_days ws_10ms_days ws_15ms_days thunder_days dew_days hoarfrost_days
#> 12             0           12            2            0        0              9
#> 11             2            1            0            0        0              3
#> 4              7            5            0            0        2              5
#> 5              0            4            0            1       17              2
#> 9              0            1            0            9       26              0
#> 6              0            0            0            9       15              0
#>    cloud_mean_mon ws_mean_mon vapor_press_mean_mon rh_mean_mon press_mean_mon
#> 12            6.4         4.6                  5.8        80.0          967.9
#> 11            6.8         2.9                  4.1        83.8          969.5
#> 4             5.5         3.6                  5.1        72.5          959.8
#> 5             4.2         3.8                  9.0        56.7          967.9
#> 9             4.4         3.1                 12.3        68.4          969.9
#> 6             5.7         2.6                 14.8        74.2          968.6
#>    slp_mean_mon rr_daytime rr_nightime
#> 12       1017.2       13.5        10.0
#> 11       1019.9       17.4        16.7
#> 4        1008.9       14.1        14.4
#> 5        1015.0        5.3         5.0
#> 9        1016.8       40.6        50.5
#> 6        1015.2       45.7        89.8
# }