Monthly IMGW meteorological data

Downloading monthly (meteorological) data from the SYNOP / CLIMATE / PRECIP stations available in the danepubliczne.imgw.pl collection

meteo_imgw_monthly(
  rank = "synop",
  year,
  status = FALSE,
  coords = FALSE,
  station = NULL,
  col_names = "short",
  ...
)

Arguments

rank

rank of the stations: "synop" (default), "climate", or "precip"

year

vector of years (e.g., 1966:2000)

status

leave the columns with measurement and observation statuses (default status = FALSE - i.e. the status columns are deleted)

coords

add coordinates of the station (logical value TRUE or FALSE)

station

name or ID of meteorological station(s). It accepts names (characters in CAPITAL LETTERS) or stations' IDs (numeric)

col_names

three types of column names possible: "short" - default, values with shorten names, "full" - full English description, "polish" - original names in the dataset

...

other parameters that may be passed to the 'shortening' function that shortens column names

Examples

# \donttest{
  monthly = meteo_imgw_monthly(rank = "climate", year = 1969)
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/klimat/k_m_d_format.txt"
#> /tmp/RtmpZvItzl/file18558b53259
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/klimat/k_m_t_format.txt"
#> /tmp/RtmpZvItzl/file1853c29adc6
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/klimat/"
#> /tmp/RtmpZvItzl/file18570ee3988
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/klimat/1966_1970/1966_1970_m_k.zip"
#> /tmp/RtmpZvItzl/file18537ca927
  head(monthly)
#>           rank        id  station   yy mm tmax_abs tmax_mean tmin_abs tmin_mean
#> 37 KLIMATYCZNA 249180010 PSZCZYNA 1969  1      9.5      -1.3    -20.0      -9.3
#> 41 KLIMATYCZNA 249180010 PSZCZYNA 1969  2      7.4       0.4    -16.6      -4.7
#> 42 KLIMATYCZNA 249180010 PSZCZYNA 1969  3     11.4       2.5    -10.6      -3.5
#> 43 KLIMATYCZNA 249180010 PSZCZYNA 1969  4     27.5      13.1     -4.4       2.0
#> 44 KLIMATYCZNA 249180010 PSZCZYNA 1969  5     30.0      21.3      2.0       9.0
#> 45 KLIMATYCZNA 249180010 PSZCZYNA 1969  6     28.9      20.9      6.1      10.5
#>    t2m_mean_mon t5cm_min rr_monthly rr_max_daily first_day_max_rr
#> 37         -5.1    -22.4       34.9         12.0               23
#> 41         -2.3    -17.5       47.2         15.9                8
#> 42         -0.8    -11.3       44.8         10.1               11
#> 43          7.1     -6.2       15.7          4.0                1
#> 44         14.8     -0.1       69.2         30.0               29
#> 45         15.4      4.4      101.2         35.0               21
#>    last_day_max_rr snowcover_max snowcover_days rain_days snow_days rh_mean_mon
#> 37              NA            10             21         4         6           0
#> 41              NA            18             16         4         7           0
#> 42              NA             2              3         3         8           0
#> 43              NA             0              0         6         4           0
#> 44              NA             0              0         9         0           0
#> 45              NA             0              0        16         0           0
#>    ws_mean_mon cloud_mean_mon
#> 37         2.1            5.1
#> 41         2.2            6.6
#> 42         2.7            5.5
#> 43         2.1            3.5
#> 44         2.0            4.1
#> 45         1.8            5.3

  # a descriptive (long) column names:
  monthly2 = meteo_imgw_monthly(rank = "synop", year = 2018,
         col_names = "full")
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/s_m_d_format.txt"
#> /tmp/RtmpZvItzl/file185766b74bc
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/s_m_t_format.txt"
#> /tmp/RtmpZvItzl/file18515b896e5
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/"
#> /tmp/RtmpZvItzl/file18531ee1102
#> [1] "https://danepubliczne.imgw.pl/data/dane_pomiarowo_obserwacyjne/dane_meteorologiczne/miesieczne/synop/2018/2018_m_s.zip"
#> /tmp/RtmpZvItzl/file18567c23bda
  head(monthly2)
#>   Station rank code        ID       Station Year Month
#> 1       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     1
#> 2       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     2
#> 3       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     3
#> 4       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     4
#> 5       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     5
#> 6       SYNOPTYCZNA 349190600 BIELSKO-BIAŁA 2018     6
#>   Absolute maximum air temperature [C] Monthly mean maximum air temperature [C]
#> 1                                 11.6                                      5.0
#> 2                                  7.2                                     -1.4
#> 3                                 16.5                                      4.8
#> 4                                 26.7                                     20.3
#> 5                                 28.0                                     21.3
#> 6                                 29.5                                     22.7
#>   Absolute minimum air temperature [C] Monthly mean minimum air temperature [C]
#> 1                                 -7.7                                     -0.4
#> 2                                -15.5                                     -6.0
#> 3                                -17.2                                     -3.0
#> 4                                 -1.0                                      8.2
#> 5                                  5.1                                     10.6
#> 6                                  8.1                                     13.5
#>   Monthly mean air temperature [C]
#> 1                              2.3
#> 2                             -3.7
#> 3                              1.0
#> 4                             14.3
#> 5                             16.2
#> 6                             17.6
#>   Minimum daily near surface air temperature [C]
#> 1                                          -10.8
#> 2                                          -23.3
#> 3                                          -21.9
#> 4                                           -1.7
#> 5                                            2.2
#> 6                                            6.1
#>   Total monthly precipitation [mm] Maximum daily preciptation [mm]
#> 1                             23.5                             6.5
#> 2                             34.1                             9.5
#> 3                             28.5                            10.1
#> 4                             10.3                             3.5
#> 5                             91.1                            19.5
#> 6                            135.5                            30.4
#>   First day with maximum precipitation Last day with maximum precipitation
#> 1                                    8                                  NA
#> 2                                    4                                  NA
#> 3                                   16                                  NA
#> 4                                   23                                  NA
#> 5                                   17                                  NA
#> 6                                    2                                  NA
#>   Total monthly insolation [hours] Maximum daily snow cover [cm]
#> 1                             58.8                             4
#> 2                             47.0                            19
#> 3                            110.7                            15
#> 4                            218.2                             0
#> 5                            247.1                             0
#> 6                            164.1                             0
#>   Days with snow cover Days with rain Days with snow Days with sleet
#> 1                    5              7             10               4
#> 2                   25              0             16               4
#> 3                   14              5              7               2
#> 4                    0              7              1               1
#> 5                    0             14              0               0
#> 6                    0             16              1               0
#>   Days with hail Days with fog Days with fogginess Days with rime
#> 1              0             5                  25              0
#> 2              0             7                  28              0
#> 3              0             2                  24              0
#> 4              0             1                   6              0
#> 5              1             1                  11              0
#> 6              0             2                  15              0
#>   Days with glaze Days with low snowstorm Days with high snowstorm
#> 1               0                       0                        0
#> 2               1                       6                        1
#> 3               0                       2                        0
#> 4               0                       0                        0
#> 5               0                       0                        0
#> 6               0                       0                        0
#>   Days with hazyness Days with wind speed above 10 ms
#> 1                  0                               12
#> 2                  2                                1
#> 3                  7                                5
#> 4                  0                                4
#> 5                  0                                1
#> 6                  0                                0
#>   Days with wind speed above 15 ms Days with thunderstorm Days with dew
#> 1                                2                      0             0
#> 2                                0                      0             0
#> 3                                0                      0             2
#> 4                                0                      1            17
#> 5                                0                      9            26
#> 6                                0                      9            15
#>   Days with hoarfrost Monthly mean total cloudiness [hPa]
#> 1                   9                                 6.4
#> 2                   3                                 6.8
#> 3                   5                                 5.5
#> 4                   2                                 4.2
#> 5                   0                                 4.4
#> 6                   0                                 5.7
#>   Monthly mean wind speed [m/s] Monthly mean vapour pressure [hPa]
#> 1                           4.6                                5.8
#> 2                           2.9                                4.1
#> 3                           3.6                                5.1
#> 4                           3.8                                9.0
#> 5                           3.1                               12.3
#> 6                           2.6                               14.8
#>   Monthly mean relative humidity [%]
#> 1                               80.0
#> 2                               83.8
#> 3                               72.5
#> 4                               56.7
#> 5                               68.4
#> 6                               74.2
#>   Monthly mean pressure at station level [hPa]
#> 1                                        967.9
#> 2                                        969.5
#> 3                                        959.8
#> 4                                        967.9
#> 5                                        969.9
#> 6                                        968.6
#>   Monthly mean sea level pressure [hPa] Total precipitation during day [mm]
#> 1                                1017.2                                13.5
#> 2                                1019.9                                17.4
#> 3                                1008.9                                14.1
#> 4                                1015.0                                 5.3
#> 5                                1016.8                                40.6
#> 6                                1015.2                                45.7
#>   Total precipitation during night [mm]
#> 1                                  10.0
#> 2                                  16.7
#> 3                                  14.4
#> 4                                   5.0
#> 5                                  50.5
#> 6                                  89.8

  # please note that station names may change over time
  # and thus 2 names are required in some cases:
  # df = meteo_imgw_monthly(rank = 'synop', year = 1991:2000,
  #         coords = TRUE, station = c("POZNAŃ","POZNAŃ-ŁAWICA"))
# }