Scrapping meteorological (Synop) data from the Ogimet webpage

Downloading hourly or daily (meteorological) data from the Synop stations available at https://www.ogimet.com/

meteo_ogimet(
  interval,
  date = c(Sys.Date() - 30, Sys.Date()),
  coords = FALSE,
  station,
  precip_split = TRUE,
  allow_failure = TRUE
)

Arguments

interval

'daily' or 'hourly' dataset to retrieve - given as character

date

start and finish date (e.g., date = c("2018-05-01", "2018-07-01")) - character or Date class object. If not provided last 30 days are used.

coords

add geographical coordinates of the station (logical value TRUE or FALSE)

station

WMO ID of meteorological station(s). Character or numeric vector

precip_split

whether to split precipitation fields into 6/12/24h

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE numeric fields (logical value TRUE (default) or FALSE); valid only for hourly time step

Value

A data.frame of measured values with columns describing the meteorological parameters (e.g. air temperature, wind speed, cloudines). Depending on the interval, at a given hour or day. Different parameters are returned for daily and hourly datasets.

station_ID - WMO station identifier

Lon - longitude

Lat - latitude

Date - date (and time) of observations

TC - air temperature at 2 metres above ground level. Values given in Celsius degrees

TdC - dew point temperature at 2 metres above ground level. Values given in Celsius degrees

TmaxC - maximum air temperature at 2 metres above ground level. Values given in Celsius degrees

TminC - minimum air temperature at 2 metres above ground level. Values given in Celsius degrees

ddd - wind direction

ffkmh - wind speed in km/h

Gustkmh - wind gust in km/h

P0hpa - air pressure at elevation of the station in hPa

PseahPa - sea level pressure in hPa

PTnd - pressure tendency in hPa

Nt - total cloud cover

Nh - cloud cover by high-level cloud fraction

HKm - height of cloud base

InsoD1 - insolation in hours

Viskm - visibility in kilometres

Snowcm - depth of snow cover in centimetres

pr6 - precicipitation totals in 6 hours

pr12 - precicipitation totals in 12 hours

pr24 - precicipitation totals in 24 hours

TemperatureCAvg - average air temperature at 2 metres above ground level. Values given in Celsius degrees

TemperatureCMax - maximum air temperature at 2 metres above ground level. Values given in Celsius degrees

TemperatureCMin - minimum air temperature at 2 metres above ground level. Values given in Celsius degrees

TdAvgC - average dew point temperature at 2 metres above ground level. Values given in Celsius degrees

HrAvg - average relative humidity. Values given in %

WindkmhDir - wind direction

WindkmhInt - wind speed in km/h

WindkmhGust - wind gust in km/h

PresslevHp - Sea level pressure in hPa

Precmm - precipitation totals in mm

TotClOct - total cloudiness in octants

lowClOct - cloudiness by low level clouds in octants

SunD1h - sunshine duration in hours

PreselevHp - atmospheric pressure measured at altitude of station in hPa

SnowDepcm - depth of snow cover in centimetres

Examples

# \donttest{
  # downloading daily data for New York - La Guardia (last 30 days by default)
  new_york = meteo_ogimet(interval = "daily", 
                          station = 72503, 
                          coords = TRUE)
#> Daily raports will be generated for 6 UTC each day. Use the >>hour<< argument to change it 
#> [1] 72503
#> [1] "https://www.ogimet.com/cgi-bin/gsynres?lang=en&ind=72503&ndays=32&ano=2022&mes=10&day=12&hora=6&ord=REV&Send=Send"
#> /tmp/RtmpRGSeu8/file18149753cb
#> [1] "https://www.ogimet.com/cgi-bin/gsynres?lang=en&ind=72503&ndays=32&ano=2022&mes=09&day=12&hora=6&ord=REV&Send=Send"
#> /tmp/RtmpRGSeu8/file18151772015
# }