Scrapping daily meteorological (Synop) data from the Ogimet webpage

Downloading daily (meteorological) data from the Synop stations available in the https://www.ogimet.com/ repository. The data are processed only if temperature or precipitation fields are present.

ogimet_daily(
  date = c(Sys.Date() - 30, Sys.Date()),
  coords = FALSE,
  station = NA,
  hour = 6,
  allow_failure = TRUE
)

Arguments

date

start and finish of date (e.g., date = c("2018-05-01","2018-07-01") ). By default last 30 days.

coords

add geographical coordinates of the station (logical value TRUE or FALSE)

station

WMO ID of meteorological station(s). Character or numeric vector

hour

time for which the daily raport is generated. Set default as hour = 6 (i.e. 6 UTC)

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

Value

data.frame with historical meteorological data for the daily summaries

Examples

# \donttest{
  # downloading daily summaries for last 30 days. station: New York - La Guardia
  new_york = ogimet_daily(station = 72503, coords = TRUE)
#> Daily raports will be generated for 6 UTC each day. Use the >>hour<< argument to change it 
#> station: 72503
#> INFO: Please note that the Ogimet has recently limited number of queries that are accepted
#>          by the server from a single IP address. Therefore, downloading more than 1 month of data
#>          for a single station requires 20 seconds pause between subsequent queries and
#>          may take a while. Thank you for your patience.
# }