Sounding data

Downloading the measurements of the vertical profile of atmosphere (also known as sounding data). Data can be retrieved using TEMP and BUFR sounding formatting. By default automatic detection of input format is used.

sounding_wyoming(
  wmo_id,
  yy,
  mm,
  dd,
  hh,
  min = 0,
  source = "AUTODETECT",
  allow_failure = TRUE
)

Source

http://weather.uwyo.edu/upperair/sounding.html

Arguments

wmo_id

international WMO station code (World Meteorological Organization ID); For Polish stations: Leba - 12120, Legionowo - 12374, Wrocław - 12425

yy

year - calendar year

mm

month - calendar month

dd

day - calendar day

hh

hour - hour of sounding; for most stations measurements are performed twice a day (i.e. at 12 and 00 UTC), sporadically 4 times a day

min

minute - minute of sounding; Default 00; Other values applies only to BUFR soundings.

source

optional; AUTODETECT, BUFR (raw; most detailed) or TEMP (FM-35) input format. By default "AUTODETECT" is used which means that the function will try to detect the format automatically. If a station use only a specific format that is known you can specify it explicitly.

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

Value

Returns two lists with values described at: weather.uwyo.edu ; The first list contains:

PRES - Pressure (hPa)

HGHT - Height (metres)

TEMP - Temperature (C)

DWPT - Dew point (C)

RELH - Relative humidity (%)

MIXR - Mixing ratio (g/kg)

DRCT - Wind direction (deg)

SPED - Wind speed (m/s)

THTA - (K)

THTE - (K)

THTV - (K)

The second list contains metadata for time of observation and station location

A list of 2 data.frames where first data frame represents parameters of upper parts o with columns describing the meteorogical parameters (e.g. temperature, air pressure) where each row represent a measurement, depending on the height. Second data.frame presents a description of the conditions under which the sounding was carried out.

Examples

# \donttest{
# download data for Station 45004 starting 1120Z 11 Jul 2021; Kowloon, HONG KONG, CHINA
# using AUTODETECT, TEMP and BUFR sounding formats

# autodect input format:
sounding_auto = sounding_wyoming(wmo_id = 45004, 
                                 yy = 2021, mm = 07, dd = 17, hh = 12)
#> https://weather.uwyo.edu/wsgi/sounding?datetime=2021-07-17%2012:00:00&id=45004&src=UNKNOWN&type=TEXT:LIST

# temp (fm35) input format:
sounding_temp = sounding_wyoming(wmo_id = 45004, 
                                 yy = 2021, mm = 07, dd = 17, hh = 12, 
                                 source = "TEMP")
#> https://weather.uwyo.edu/wsgi/sounding?datetime=2021-07-17%2012:00:00&id=45004&src=FM35&type=TEXT:LIST
# bufr input format:  
sounding_bufr = sounding_wyoming(wmo_id = 45004, 
                                 yy = 2021, mm = 07, dd = 17, hh = 12, min = 00, 
                                 source = "BUFR")
#> https://weather.uwyo.edu/wsgi/sounding?datetime=2021-07-17%2012:00:00&id=45004&src=BUFR&type=TEXT:LIST
# }