Sounding data

Downloading the measurements of the vertical profile of atmosphere (also known as sounding data). Data can be retrieved using TEMP and BUFR sounding formatting.

sounding_wyoming(
  wmo_id,
  yy,
  mm,
  dd,
  hh,
  min = 0,
  bufr = FALSE,
  allow_failure = TRUE
)

Source

http://weather.uwyo.edu/upperair/sounding.html

Arguments

wmo_id

international WMO station code (World Meteorological Organization ID); For Polish stations: Łeba - 12120, Legionowo - 12374, Wrocław- 12425

yy

year - single number

mm

month - single number denoting month

dd

day - single number denoting day

hh

hour - single number denoting initial hour of sounding; for most stations this measurement is done twice a day (i.e. at 12 and 00 UTC), sporadically 4 times a day

min

minute - single number denoting initial minute of sounding; applies only to BUFR soundings.

bufr

BUFR or TEMP sounding to be decoded. By default TEMP is used. For BUFR soundings use bufr = TRUE

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

Value

Returns two lists with values described at: weather.uwyo.edu ; The first list contains:

PRES - Pressure (hPa)

HGHT - Height (metres)

TEMP - Temperature (C)

DWPT - Dew point (C)

RELH - Relative humidity (%)

MIXR - Mixing ratio (g/kg)

DRCT - Wind direction (deg)

SKNT - Wind speed (knots)

THTA = (K)

THTE = (K)

THTV = (K)

The second list contains metadata and calculated thermodynamic / atmospheric instability indices (for TEMP soundings only)

A list of 2 data.frames where first data frame represents parameters of upper parts o with columns describing the meteorogical parameters (e.g. temperature, air pressure) where each row represent a measurement, depending on the height. Second data.frame presents a description of the conditions under which the sounding was carried out.

Examples

# \donttest{
##############################################################################
# download data for Station 45004 starting 1120Z 11 Jul 2021; Kowloon, HONG KONG, CHINA
# using TEMP and BUFR sounding formats
##############################################################################
  TEMP = sounding_wyoming(wmo_id = 45004, yy = 2021, mm = 07, dd = 17, 
                          hh = 12, min = 00)
#> [1] "http://weather.uwyo.edu/cgi-bin/sounding?TYPE=TEXT%3ALIST&YEAR=2021&MONTH=07&FROM=1712&TO=1712&STNM=45004"
#> /tmp/RtmpRYa28L/file1824423ad1d
  #head(TEMP[[1]])
  
  BUFR = sounding_wyoming(wmo_id = 45004, yy = 2021, mm = 07, dd = 17, 
                          hh = 12, min = 00, bufr = TRUE)
#> [1] "http://weather.uwyo.edu/cgi-bin/bufrraob.py?datetime=2021-07-17+12:00:00&id=45004&type=TEXT:LIST"
#> /tmp/RtmpRYa28L/file1827b85b529
  #head(BUFR[[1]])

##############################################################################
### example with a random date to download sounding from LEBA, PL station: ###
##############################################################################

  profile = sounding_wyoming(wmo_id = 12120, 
                             yy = sample(2000:2019,1),
                             mm = sample(1:12,1), 
                             dd = sample(1:20,1), 
                             hh = 0)
#> [1] "http://weather.uwyo.edu/cgi-bin/sounding?TYPE=TEXT%3ALIST&YEAR=2005&MONTH=11&FROM=1200&TO=1200&STNM=12120"
#> /tmp/RtmpRYa28L/file182450910db
  # head(profile)
  # plot(profile[[1]]$HGHT, profile[[1]]$PRES, type = 'l')
# }