Sounding data

Downloading the measurements of the vertical profile of atmosphere (also known as sounding data). Data can be retrieved using TEMP and BUFR sounding formatting.

sounding_wyoming(
  wmo_id,
  yy,
  mm,
  dd,
  hh,
  min = 0,
  bufr = FALSE,
  allow_failure = TRUE
)

Source

http://weather.uwyo.edu/upperair/sounding.html

Arguments

wmo_id

international WMO station code (World Meteorological Organization ID); For Polish stations: Leba - 12120, Legionowo - 12374, Wrocław- 12425

yy

year - single number

mm

month - single number denoting month

dd

day - single number denoting day

hh

hour - single number denoting initial hour of sounding; for most stations this measurement is done twice a day (i.e. at 12 and 00 UTC), sporadically 4 times a day

min

minute - single number denoting initial minute of sounding; applies only to BUFR soundings.

bufr

BUFR or TEMP sounding to be decoded. By default TEMP is used. For BUFR soundings use bufr = TRUE

allow_failure

logical - whether to proceed or stop on failure. By default set to TRUE (i.e. don't stop on error). For debugging purposes change to FALSE

Value

Returns two lists with values described at: weather.uwyo.edu ; The first list contains:

PRES - Pressure (hPa)

HGHT - Height (metres)

TEMP - Temperature (C)

DWPT - Dew point (C)

RELH - Relative humidity (%)

MIXR - Mixing ratio (g/kg)

DRCT - Wind direction (deg)

SKNT - Wind speed (knots)

THTA = (K)

THTE = (K)

THTV = (K)

The second list contains metadata and calculated thermodynamic / atmospheric instability indices (for TEMP soundings only)

A list of 2 data.frames where first data frame represents parameters of upper parts o with columns describing the meteorogical parameters (e.g. temperature, air pressure) where each row represent a measurement, depending on the height. Second data.frame presents a description of the conditions under which the sounding was carried out.

Examples

# \donttest{
##############################################################################
# download data for Station 45004 starting 1120Z 11 Jul 2021; Kowloon, HONG KONG, CHINA
# using TEMP and BUFR sounding formats
##############################################################################
  TEMP = sounding_wyoming(wmo_id = 45004, yy = 2021, mm = 07, dd = 17, 
                          hh = 12, min = 00)
#> /tmp/Rtmpaz3JGC/file1ec9bb65c9e
  head(TEMP[[1]])
#>   PRES HGHT TEMP DWPT RELH  MIXR DRCT SKNT  THTA  THTE  THTV
#> 1 1000   50   NA   NA   NA    NA   NA   NA    NA    NA    NA
#> 2  998   66 27.6 23.6   79 18.78    0    0 300.9 356.4 304.3
#> 3  975  272 25.9 22.8   83 18.34  105   10 301.3 355.5 304.6
#> 4  925  737 22.2 21.1   93 17.36  120   19 302.0 353.4 305.1
#> 5  921  775 22.1 20.9   93 17.22  120   21 302.2 353.3 305.4
#> 6  907  909 21.6 20.2   92 16.73  114   20 303.1 352.9 306.1
  
  BUFR = sounding_wyoming(wmo_id = 45004, yy = 2021, mm = 07, dd = 17, 
                          hh = 12, min = 00, bufr = TRUE)
#> /tmp/Rtmpaz3JGC/file1ec961835834
  head(BUFR[[1]])
#>     PRES HGHT TEMP DWPT RELH  MIXR DRCT SKNT  THTA  THTE  THTV
#> 1 1000.0   50   NA   NA   NA    NA   NA   NA    NA    NA    NA
#> 2  998.2   66 27.7 23.7   79 18.85   27  0.4 301.0 356.7 304.4
#> 3  996.1   84 27.7 23.8   79 19.00  120  4.7 301.2 357.4 304.6
#> 4  995.1   93 27.7 23.8   79 18.96  120  5.6 301.3 357.4 304.7
#> 5  993.9  104 27.6 23.6   79 18.76  120  6.0 301.3 356.8 304.7
#> 6  992.7  115 27.5 23.5   79 18.69  120  6.4 301.3 356.6 304.6

##############################################################################
### example with a random date to download sounding from LEBA, PL station: ###
##############################################################################

  profile = sounding_wyoming(wmo_id = 12120, 
                             yy = sample(2000:2019,1),
                             mm = sample(1:12,1), 
                             dd = sample(1:20,1), 
                             hh = 0)
#> /tmp/Rtmpaz3JGC/file1ec95ba2ef21
  # plot(profile[[1]]$HGHT, profile[[1]]$PRES, type = 'l')
# }